-apply readdirpp fix
git-svn-id: https://svn.cern.ch/reps/lcgdm/davix/trunk@6230 4525493e-7705-40b1-a816-d608a930855b

+++ src/fileproperties.h
@@ -10,7 +10,7 @@ struct FileProperties
 {
 
     Glib::ustring  filename;
-    long      req_status; /* status code of the request associated ( ex: http 200) */
+    unsigned int  req_status; /* status code of the request associated ( ex: http 200) */
 
     nlink_t   nlink;
     uid_t     uid;    /* unix user id */
